FHAC-94 Refactor of ping service, wsdl output configurable, and clean up of blockui
