+++ Product/Production/Adapters/General/CONNECTAdminGUI/src/main/webapp/connectionManager.xhtml
@@ -132,7 +132,7 @@
                         </p:row>
                     </p:panelGrid>
                     
-                    <p:blockUI trigger="ping-command" block="endpointTable">
+                    <p:blockUI trigger="ping-command" block="orgPanel">
                             LOADING...<br/>
                             <p:graphicImage library="images" name="loading-bar-gif.gif" />
                         </p:blockUI>
